Refactor auth_token into multiple files
Also, added documentation comments, and cleaned up the tests by using parameterization
